typescript migration
+ gruntfile integration with ts
+ build file (inc. implementation, test and assets)
