+++ .gitignore
@@ -1,4 +1,7 @@
 node_modules
+package-lock.json
 .vagrant
 _test
 .DS_Store
+build
+.tscache
