+++ package.json
@@ -2,9 +2,10 @@
   "name": "mysql-wrap-production",
   "version": "0.11.3",
   "description": "Mysql interface and helper functions, wrapping node-mysql",
-  "main": "src/mysql-wrap.js",
+  "main": "build/mysql-wrap.js",
   "scripts": {
-    "test": "grunt test"
+    "test": "grunt test",
+    "build": "grunt test"
   },
   "repository": {
     "type": "git",
@@ -23,15 +24,29 @@
     "url": "https://github.com/tourmalinelabs/mysql-wrap-production/issues"
   },
   "devDependencies": {
+    "@types/jest": "^26.0.24",
+    "@types/mocha": "^9.0.0",
+    "@types/node": "^16.4.13",
     "chai": "^4.3.4",
     "grunt": "^1.4.1",
+    "grunt-contrib-copy": "^1.0.0",
     "grunt-mocha-test": "^0.13.3",
     "mocha": "^9.0.3",
-    "mysql": "^2.18.1"
+    "mysql": "^2.18.1",
+    "typescript": "^4.3.5"
   },
   "dependencies": {
+    "@types/lodash": "^4.14.172",
+    "@typescript-eslint/eslint-plugin": "^4.29.0",
+    "@typescript-eslint/parser": "^4.29.0",
     "bluebird": "^3.7.2",
+    "del": "^6.0.0",
+    "eslint": "^7.32.0",
+    "eslint-config-prettier": "^8.3.0",
+    "eslint-plugin-prettier": "^3.4.0",
+    "grunt-ts": "^6.0.0-beta.22",
     "lodash": "^4.17.21",
+    "prettier": "^2.3.2",
     "squel": "^5.13.0"
   }
 }
