+++ src/config.json
@@ -2,7 +2,7 @@
     "mysql": {
         "host": "127.0.0.1",
         "user": "root",
-        "password": "password",
+        "password": "secret",
         "database": "mysql-test"
     }
 }
